requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# FastAPI Stuff
fastapi
uvicorn

# LLM model
llama-cpp-python